Add glaze_error_category and make_error_code for use with std::error_code
